more tweaks

+++ api/controllers/directors/remove-hours.js
@@ -17,15 +17,16 @@ module.exports = {
     fn: async function (inputs, exits) {
         sails.log.debug('Controller directors/remove-hours called.');
         try {
-            
+
             // Delete records from director calendar
             var records = await Directorhours.destroy({ID: inputs.ID}).fetch();
-            
+
             // Also delete records from timesheets
             var IDs = [];
             records.map((record) => IDs.push(record.unique));
-            await Timesheet.destroy({unique: IDs}).fetch();
-            
+            if (IDs.length > 0)
+                await Timesheet.destroy({unique: IDs}).fetch();
+
             return exits.success();
         } catch (e) {
             return exits.error(e);
